+++ packages/svelte-maplibre-storymap/src/lib/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './StoryMapConfig.js';
 export * from './StoryMapChapter.js';
+export * from './StoryMapTemplate.js';