+++ sites/geohub/src/routes/(app)/tools/+page.ts
@@ -1,13 +1,14 @@
 import type { PageLoad } from './$types';
 
 export const load: PageLoad = async ({ data }) => {
-	const { algorithms } = data;
+	const { algorithms, datasets } = data;
 	const title = 'Tools | GeoHub';
 	const content = 'Tools';
 
 	return {
 		title,
 		content,
-		algorithms
+		algorithms,
+		datasets
 	};
 };